+++ src/pyobo/database/sql/models.py
@@ -43,7 +43,7 @@ class Resource(Base):
     """A resource."""
 
     __tablename__ = 'resource'
-    prefix = Column(String, nullable=False, unique=True, index=True, primary_key=True)
+    prefix = Column(String(32), nullable=False, unique=True, index=True, primary_key=True)
 
     name = Column(String, nullable=False, unique=True, index=True)
     pattern = Column(String, nullable=False)
@@ -57,7 +57,7 @@ class Reference(Base):
     """Represent a CURIE and label."""
 
     __tablename__ = 'reference'
-    prefix = Column(String, ForeignKey(f'{Resource.__tablename__}.prefix'), primary_key=True)
+    prefix = Column(String(32), ForeignKey(f'{Resource.__tablename__}.prefix'), primary_key=True)
     identifier = Column(String, primary_key=True)
 
     name = Column(String, index=True, nullable=True)
@@ -80,7 +80,7 @@ class Synonym(Base):
     __tablename__ = 'synonym'
     id = Column(Integer, primary_key=True)
 
-    prefix = Column(String, ForeignKey(f'{Resource.__tablename__}.prefix'))
+    prefix = Column(String(32), ForeignKey(f'{Resource.__tablename__}.prefix'))
     identifier = Column(String)
     name = Column(String, index=True)
 
@@ -100,7 +100,7 @@ class Alt(Base):
     """Represents an alternate identifier relationship."""
 
     __tablename__ = 'alt'
-    prefix = Column(String, ForeignKey(f'{Resource.__tablename__}.prefix'), primary_key=True)
+    prefix = Column(String(32), ForeignKey(f'{Resource.__tablename__}.prefix'), primary_key=True)
     alt = Column(String, primary_key=True)
 
     identifier = Column(String, index=True)
@@ -118,10 +118,10 @@ class Xref(Base):
     __tablename__ = 'xref'
     id = Column(Integer, primary_key=True)
 
-    prefix = Column(String, ForeignKey(f'{Resource.__tablename__}.prefix'))
+    prefix = Column(String(32), ForeignKey(f'{Resource.__tablename__}.prefix'))
     identifier = Column(String)
 
-    xref_prefix = Column(String)
+    xref_prefix = Column(String(32))
     xref_identifier = Column(String)
 
     source = Column(Text, index=True)
