+++ src/pyobo/sources/conso.py
@@ -24,6 +24,7 @@ def get_obo() -> Obo:
         ontology=PREFIX,
         name='Curation of Neurodegeneration Supporting Ontology',
         iter_terms=iter_terms,
+        auto_generated_by=f'bio2obo:{PREFIX}',
     )
