+++ src/pyobo/sources/drugbank.py
@@ -28,6 +28,7 @@ def get_obo() -> Obo:
         ontology=PREFIX,
         name='DrugBank',
         iter_terms=iter_terms,
+        auto_generated_by=f'bio2obo:{PREFIX}',
     )
